Extract proxy filtering module for strategy opt-in
