+++ lib/rack_request_ip_strategies/base.rb
@@ -11,18 +11,6 @@ module RackRequestIPStrategies
 
     private
 
-    def filter_proxies_from(header)
-      filter_proxies(ips_from(header))
-    end
-
-    def filter_proxies(ips)
-      ips.reject { |ip| trusted_proxy?(ip) }
-    end
-
-    def trusted_proxy?(ip)
-      TrustedProxyDetector.trusted_proxy?(ip, @config.trusted_proxies)
-    end
-
     def ips_from(header)
       ips = @env[header] ? @env[header].strip.split(/[,\s]+/) : []
       ips.select { |ip| valid_ip?(ip) }
