+++ lib/rack_request_ip_strategies/proxy_filtering.rb
@@ -1,0 +1,15 @@
+module RackRequestIPStrategies
+  module ProxyFiltering
+    def filter_proxies_from(header)
+      filter_proxies(ips_from(header))
+    end
+
+    def filter_proxies(ips)
+      ips.reject { |ip| trusted_proxy?(ip) }
+    end
+
+    def trusted_proxy?(ip)
+      TrustedProxyDetector.trusted_proxy?(ip, @config.trusted_proxies)
+    end
+  end
+end
