+++ lib/rack_request_ip_strategies/x_forwarded_for.rb
@@ -1,5 +1,7 @@
 module RackRequestIPStrategies
   class XForwardedFor < Base
+    include ProxyFiltering
+
     def calculate
       filter_proxies_from('HTTP_X_FORWARDED_FOR').last
     end
